Fix merge mistakes
