+++ test/functional/apps/console/_misc_console_behavior.ts
@@ -6,6 +6,9 @@
  * Side Public License, v 1.
  */
 import expect from '@kbn/expect';
+import { REPO_ROOT } from '@kbn/repo-info';
+import { existsSync, readFileSync, writeFileSync, unlinkSync } from 'fs';
+import { resolve } from 'path';
 import { FtrProviderContext } from '../../ftr_provider_context';
 
 export default function ({ getService, getPageObjects }: FtrProviderContext) {
@@ -205,6 +208,49 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
           expect(actualResponse).to.contain('parsing_exception');
           expect(await PageObjects.console.hasSuccessBadge()).to.be(false);
         });
+      });
+    });
+
+    describe('import/export file', () => {
+      it('can import file into the editor', async () => {
+        await PageObjects.console.clearEditorText();
+
+        // Create input file path
+        const filePath = resolve(REPO_ROOT, `target/functional-tests/downloads/console_import`);
+        writeFileSync(filePath, 'GET _search', 'utf8');
+
+        // Set file to upload and wait for the editor to be updated
+        await PageObjects.console.setFileToUpload(filePath);
+        await PageObjects.common.sleep(1000);
+
+        await retry.try(async () => {
+          const request = await PageObjects.console.getEditorText();
+          expect(request).to.be.eql('GET _search');
+        });
+
+        // Clean up input file
+        unlinkSync(filePath);
+      });
+
+      it('can export input as file', async () => {
+        await PageObjects.console.enterText('GET _search');
+        await PageObjects.console.clickExportButton();
+
+        // Wait for download to trigger
+        await PageObjects.common.sleep(1000);
+
+        const downloadPath = resolve(REPO_ROOT, `target/functional-tests/downloads/console_export`);
+        await retry.try(async () => {
+          const fileExists = existsSync(downloadPath);
+          expect(fileExists).to.be(true);
+        });
+
+        // Verify the downloaded file content
+        const fileContent = readFileSync(downloadPath, 'utf8');
+        expect(fileContent).to.be('GET _search');
+
+        // Clean up downloaded file
+        unlinkSync(downloadPath);
       });
     });
   });
